Use cast instead of pattern matching in ByteArray
There is a concern raised by @armanbilge about using an implementation
detail of the Scala Native array implementation (the fact that `Array[Byte]`
are implemented in term of
`scalanative.runtime.ByteArray`
Since relying on this implementation detail is done already in the
library in `MessageDigest`, I'm continuing to rely on it here.
If a future version of Scala Native changes this we will need to update the
library to something different.